Accident Images Analysis Dataset
This dataset is provided for classification approaches on accident images. It includes three folders:
•	Accident –Detection
•	Vehicles-in-Accidents
•	Accident-Severity
The first folder: Accident –Detection
The aim of this dataset is to detect the occurrence of accidents, by image processing. It includes two subfolders with labels "without-accident" and "with-accident". 
	Folder 1 includes 2500 images with label "without-accident".
	Folder 2 includes 2398 images with label "with-accident" 
The second folder: Vehicles-in-Accidents
The aim of this dataset is to classify the vehicles involved in an accident, by image processing. It includes three subfolders with labels "light vehicle", "heavy vehicle" and "motorcycle".
	Folder 1 includes 892 images with label "light vehicle".
	Folder 2 includes 876 images with label "heavy vehicle".
	Folder 3 includes 868 images with label "motorcycle".
The third folder: Accident-Severity
The aim of this dataset is to classify the severity of accidents, by image processing. It includes three subfolders with labels " low dangerous", " medium dangerous" and " high dangerous".
	Folder 1 includes 118 images with label "low dangerous".
	Folder 2 includes 1603 images with label "medium dangerous".
	Folder 3 includes 1225 images with label "high dangerous".
+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
In case of using this dataset for any research and/or application, please refer to the following paper:
Ali Pashaei, Mehdi Ghatee, Hedieh Sajedi, Convolution Neural Network Joint with Mixture of Extreme Learning Machines for Feature Extraction and Classification of Accident Images,
Journal of Real-Time Image Processing, 2019.
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++